+++ crates/scheduler/src/applet/store.rs
@@ -60,4 +60,15 @@ pub trait MemoryApi {
     fn from_bytes_mut<T: NoUninit + AnyBitPattern>(&self, ptr: u32) -> Result<&mut T, Trap> {
         Ok(bytemuck::from_bytes_mut(self.get_mut(ptr, core::mem::size_of::<T>() as u32)?))
     }
+
+    fn alloc_copy(&mut self, ptr_ptr: u32, len_ptr: Option<u32>, data: &[u8]) -> Result<(), Trap> {
+        let len = data.len() as u32;
+        let ptr = self.alloc(len, 1)?;
+        self.get_mut(ptr, len)?.copy_from_slice(data);
+        self.get_mut(ptr_ptr, 4)?.copy_from_slice(&ptr.to_le_bytes());
+        if let Some(len_ptr) = len_ptr {
+            self.get_mut(len_ptr, 4)?.copy_from_slice(&len.to_le_bytes());
+        }
+        Ok(())
+    }
 }