+++ crates/wire/test.sh
@@ -20,4 +20,5 @@ set -e
 test_helper
 
 cargo test --lib --features=schema
+cargo miri test --lib --features=schema yoke
 cargo check --lib --target=thumbv7em-none-eabi